refactor: rename web/svelte to web/ui and web/routes to web/api
