+++ .github/workflows/ci.yml
@@ -27,7 +27,7 @@ jobs:
           cache: npm
       - run: npm ci
       - run: npx tsc --noEmit
-      - run: npx svelte-check --workspace src/web/svelte --fail-on-warnings false
+      - run: npx svelte-check --workspace src/web/ui --fail-on-warnings false
       - run: npx tsc --noEmit -p templates/_base/tsconfig.json
 
   test:
