+++ lefthook.yml
@@ -9,7 +9,7 @@ pre-commit:
     typecheck-templates:
       run: npx tsc --noEmit -p templates/_base/tsconfig.json
     typecheck-frontend:
-      run: npx svelte-check --tsconfig src/web/svelte/tsconfig.json
+      run: npx svelte-check --tsconfig src/web/ui/tsconfig.json
 
 pre-push:
   commands:
