+++ package.json
@@ -29,8 +29,8 @@
   "scripts": {
     "dev": "tsx src/cli.ts",
     "build": "tsup && npm run build:web",
-    "build:web": "vite build --config src/web/svelte/vite.config.ts",
-    "dev:web": "vite --config src/web/svelte/vite.config.ts",
+    "build:web": "vite build --config src/web/ui/vite.config.ts",
+    "dev:web": "vite --config src/web/ui/vite.config.ts",
     "test": "node --import tsx --import ./test/setup.ts --test --test-force-exit --test-concurrency=1 test/*.test.ts",
     "lint": "biome check src/ test/",
     "lint:fix": "biome check --write src/ test/",
@@ -38,7 +38,7 @@
     "typecheck": "tsc --noEmit",
     "typecheck:templates": "tsc --noEmit -p templates/claude/tsconfig.json && tsc --noEmit -p templates/pi/tsconfig.json",
     "lint:templates": "biome check templates/",
-    "typecheck:web": "svelte-check --tsconfig src/web/svelte/tsconfig.json",
+    "typecheck:web": "svelte-check --tsconfig src/web/ui/tsconfig.json",
     "prepare": "lefthook install",
     "prepublishOnly": "npm run build",
     "db:generate": "drizzle-kit generate",
