+++ src/web/app.ts
@@ -4,23 +4,23 @@ import { csrf } from "hono/csrf";
 import { HTTPException } from "hono/http-exception";
 import log from "../lib/logger.js";
 import { checkForUpdateCached, getCurrentVersion } from "../lib/update-check.js";
+import auth from "./api/auth.js";
+import channels from "./api/channels.js";
+import connectors from "./api/connectors.js";
+import envRoutes, { sharedEnvApp } from "./api/env.js";
+import files from "./api/files.js";
+import logs from "./api/logs.js";
+import minds from "./api/minds.js";
+import pages from "./api/pages.js";
+import schedules from "./api/schedules.js";
+import system from "./api/system.js";
+import typing from "./api/typing.js";
+import update from "./api/update.js";
+import variants from "./api/variants.js";
+import chat from "./api/volute/chat.js";
+import conversations from "./api/volute/conversations.js";
+import userConversations from "./api/volute/user-conversations.js";
 import { authMiddleware } from "./middleware/auth.js";
-import auth from "./routes/auth.js";
-import channels from "./routes/channels.js";
-import connectors from "./routes/connectors.js";
-import envRoutes, { sharedEnvApp } from "./routes/env.js";
-import files from "./routes/files.js";
-import logs from "./routes/logs.js";
-import minds from "./routes/minds.js";
-import pages from "./routes/pages.js";
-import schedules from "./routes/schedules.js";
-import system from "./routes/system.js";
-import typing from "./routes/typing.js";
-import update from "./routes/update.js";
-import variants from "./routes/variants.js";
-import chat from "./routes/volute/chat.js";
-import conversations from "./routes/volute/conversations.js";
-import userConversations from "./routes/volute/user-conversations.js";
 
 const app = new Hono();
