+++ src/web/ui/src/components/FileEditor.svelte
