+++ src/web/ui/src/components/GroupModal.svelte
