+++ src/web/ui/src/components/LogViewer.svelte
