+++ src/web/ui/src/components/MindCard.svelte
