+++ src/web/ui/src/components/MindPickerModal.svelte
