+++ src/web/ui/src/components/SeedModal.svelte
