+++ src/web/ui/src/components/StatusBadge.svelte
