+++ src/web/ui/src/components/SystemLogs.svelte
