+++ src/web/ui/src/components/UpdateBanner.svelte
