+++ src/web/ui/src/components/VariantList.svelte
