+++ src/web/ui/src/pages/Dashboard.svelte
