+++ src/web/ui/src/pages/MindDetail.svelte
