+++ src/web/ui/svelte.config.js
